requirements.txt
alembic==1.5.8
appdirs==1.4.4
attrs==20.3.0
backcall==0.2.0
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
darkskylib==0.3.91
decorator==5.0.7
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
ipython==7.22.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
Mako==1.1.4
mapq==0.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
numpy==1.20.2
packaging==20.9
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pluggy==0.13.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
py==1.10.0
Pygments==2.8.1
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.1
python-dotenv==0.17.0
python-editor==1.0.4
pytz==2021.1
regex==2021.4.4
requests==2.25.1
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.4.5
toml==0.10.2
traitlets==5.0.5
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1